github-app/requirements.txt
# GitHub App dependencies
PyGithub>=2.1.0
pydantic>=2.0.0
pyyaml>=6.0
cryptography>=42.0.0
PyJWT>=2.8.0

# Server
fastapi>=0.100.0
uvicorn>=0.20.0
